+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* 2026-09-07 - AB#210: Reset persistent keys on each request.
 * 2026-09-07 - AB#210: Prefer flow items dictionary over individual members.
 * 2026-09-04 - AB#224: Latest NuGet for LogicBuilder.App.Spa.Business.
 * 2026-09-04 - AB#224: Once again with coverage settings.